Updated responses to publicly use variables
