--- Sources/Frame-iOS/Networking/NetworkingConstants.swift
@@ -1,10 +1,0 @@
-//
-//  NetworkingConstants.swift
-//  Frame-iOS
-//
-//  Created by Frame Payments on 9/26/24.
-//
-
-class NetworkingConstants {
-    static let mainAPIURL: String = "https://api.framepayments.com"
-}
